some mpi improvements (#139)
* some more communication routines

* removing IndexType2 as it need to be int64_t all the time

* clang format

* removing deprecated global vector routine

* matrix sizes should always be int64_t, no matter if local or global

* adjusted death tests

* adjusting remaining API exposed sizes to be int64_t

* version bump

* adjusted cg-rsamg example

* omp typo bug in host vector

* adjusting solvers for int64_t sizes

* allocate functions should also take int64_t sizes

* removing unrecognized compiler options ; adding option for different int compile modes

* int64 sizes for hip allocate functions

* routine to convert int64_t to double

* adjusting headers for updated types.hpp location

* clang format screwed .txt file up ; reverting

* should be <= numeric limits

* no need to check for int64 overflow

* removing redundant asserts

+++ src/solvers/iter_ctrl.cpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2018-2021 Advanced Micro Devices, Inc. All rights Reserved.
+ * Copyright (C) 2018-2023 Advanced Micro Devices, Inc. All rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -182,7 +182,7 @@ namespace rocalution
         return this->current_res_;
     }
 
-    int IterationControl::GetAmaxResidualIndex(void) const
+    int64_t IterationControl::GetAmaxResidualIndex(void) const
     {
         return this->current_index_;
     }
@@ -247,7 +247,7 @@ namespace rocalution
         return false;
     }
 
-    bool IterationControl::CheckResidual(double res, int index)
+    bool IterationControl::CheckResidual(double res, int64_t index)
     {
         this->current_index_ = index;
         return this->CheckResidual(res);
