+++ src/solvers/preconditioners/preconditioner_multicolored_ilu.hpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2018-2021 Advanced Micro Devices, Inc. All rights Reserved.
+ * Copyright (C) 2018-2023 Advanced Micro Devices, Inc. All rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -91,7 +91,7 @@ namespace rocalution
         /** \brief Perform factorization with levels or not */
         bool level_;
         /** \brief Number of non-zeros */
-        int nnz_;
+        int64_t nnz_;
     };
 
 } // namespace rocalution
