+++ src/solvers/solver.cpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2018-2021 Advanced Micro Devices, Inc. All rights Reserved.
+ * Copyright (C) 2018-2023 Advanced Micro Devices, Inc. All rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -288,9 +288,9 @@ namespace rocalution
     }
 
     template <class OperatorType, class VectorType, typename ValueType>
-    int IterativeLinearSolver<OperatorType, VectorType, ValueType>::GetAmaxResidualIndex(void)
+    int64_t IterativeLinearSolver<OperatorType, VectorType, ValueType>::GetAmaxResidualIndex(void)
     {
-        int ind = this->iter_ctrl_.GetAmaxResidualIndex();
+        int64_t ind = this->iter_ctrl_.GetAmaxResidualIndex();
         log_debug(this, "IterativeLinearSolver::GetAmaxResidualIndex()", ind);
 
         if(this->res_norm_type_ != 3)
