+++ src/solvers/solver.hpp
@@ -1,5 +1,5 @@
 /* ************************************************************************
- * Copyright (C) 2018-2021 Advanced Micro Devices, Inc. All rights Reserved.
+ * Copyright (C) 2018-2023 Advanced Micro Devices, Inc. All rights Reserved.
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -298,7 +298,7 @@ namespace rocalution
       * \f$L_\infty\f$ norm
       */
         ROCALUTION_EXPORT
-        virtual int GetAmaxResidualIndex(void);
+        virtual int64_t GetAmaxResidualIndex(void);
 
     protected:
         // Iteration control (monitor)
@@ -314,7 +314,7 @@ namespace rocalution
         int res_norm_type_;
 
         /** \brief Absolute maximum index of residual vector when using \f$L_\infty\f$ */
-        int index_;
+        int64_t index_;
 
         /** \brief Computes the vector norm */
         ValueType Norm_(const VectorType& vec);
